Fix the description for EPEL repos
The provided settings for the EPEL repos were using the "name" key for
the repository description, but this maps to the Puppet namevar, which
is the repo ID (aka the name of the file in /etc/yum.repos.d). They
should instead be using the key "descr" which is what maps to the "name"
value in the repo settings.

Fixes issue #53